Update mlFilters limits
