+++ src/components/TestingModelGridView.tsx
@@ -19,7 +19,6 @@ import { FormattedMessage, useIntl } from "react-intl";
 import { useConnectActions } from "../connect-actions-hooks";
 import { usePrediction } from "../hooks/ml-hooks";
 import { useProject } from "../hooks/project-hooks";
-import { mlSettings } from "../ml";
 import { getMakeCodeLang } from "../settings";
 import { useSettings, useStore } from "../store";
 import { tourElClassname } from "../tours";
@@ -31,6 +30,7 @@ import HeadingGrid from "./HeadingGrid";
 import IncompatibleEditorDevice from "./IncompatibleEditorDevice";
 import LiveGraphPanel from "./LiveGraphPanel";
 import MoreMenuButton from "./MoreMenuButton";
+import { mlSettings } from "../mlConfig";
 
 const gridCommonProps: Partial<GridProps> = {
   gridTemplateColumns: "290px 360px 40px auto",
