+++ src/store.ts
@@ -3,7 +3,9 @@ import * as tf from "@tensorflow/tfjs";
 import { create } from "zustand";
 import { devtools, persist } from "zustand/middleware";
 import { useShallow } from "zustand/react/shallow";
+import { deployment } from "./deployment";
 import { flags } from "./flags";
+import { Logging } from "./logging/logging";
 import {
   filenames,
   generateCustomFiles,
@@ -25,10 +27,8 @@ import {
   TrainModelDialogStage,
 } from "./model";
 import { defaultSettings, Settings } from "./settings";
-import { defaultIcons, MakeCodeIcon } from "./utils/icons";
-import { deployment } from "./deployment";
-import { Logging } from "./logging/logging";
 import { getTotalNumSamples } from "./utils/gestures";
+import { defaultIcons, MakeCodeIcon } from "./utils/icons";
 
 export const modelUrl = "indexeddb://micro:bit-ai-creator-model";
